Fixed Bugs and Problems
I fixed a bug that occurred with "sv-id". The bot only checked whether the server has a license, if not it says that the server does not exist. That was fixed.
The problem that one had in "sv-id" when entering letters in the ServerID led to errors in the console. In the future, a message will now be issued.