Updated to 0.4.4
A Bug was fixed that was caused when trying to kick the Bot with "kick @bot#1234". Please remove the Bot by right clicking on the 
Bot --> Kick.
The bot no longer responds to tag commands with the prefix.
He still responds to markings without commands.
The command "Hallo" is no longer triggered with the prefix. Use "Hallo @Bot#1234" to execute the command.
The command "set-patch" has been renamed to "set-news".